# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pro WhatsApp API
  version: 1.0.0
  description: |
    REST API لنظام إدارة عملاء ومحادثات WhatsApp.

    ## طرق ربط WhatsApp
    - **web** (افتراضي): جلسة WhatsApp Web عبر جسر Node.js (Baileys) — رقم واحد وعدة موظفين.
    - **cloud**: Meta WhatsApp Cloud API الرسمي.

    ## جسر WhatsApp Web
    شغّل خدمة `whatsapp-bridge` واضبط:
    ```
    WHATSAPP_BRIDGE_URL=http://127.0.0.1:3001
    WHATSAPP_BRIDGE_SECRET=shared-secret
    ```

    ## المصادقة
    بعد `POST /auth/login` أرسل التوكن في كل طلب:
    ```
    Authorization: Bearer {token}
    ```

    ## اللغة
    ```
    Accept-Language: ar
    Accept-Language: en
    ```

    ## حدود الطلبات
    | المسار | الحد |
    |--------|------|
    | عام | 120 طلب / دقيقة |
    | تسجيل الدخول | 5 طلبات / دقيقة |
    | Webhooks | 180 طلب / دقيقة |

    ## حسابات التجربة
    | الدور | البريد | كلمة المرور |
    |--------|--------|-------------|
    | Admin | admin@example.com | password |
    | Employee | ahmed@example.com | password |
  contact:
    name: Pro WhatsApp
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: تسجيل الخروج من كل الأجهزة
      operationId: logoutAll
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /auth/profile:
    put:
      tags: [Auth]
      summary: تحديث الملف الشخصي
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: الملف بعد التحديث
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/sessions/{token}:
    delete:
      tags: [Auth]
      summary: إلغاء جلسة جهاز
      operationId: revokeSession
      parameters:
        - $ref: '#/components/parameters/TokenId'
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /conversations/{conversation}/messages:
    get:
      tags: [Messages]
      summary: رسائل المحادثة
      description: الأحدث أولاً. يدعم البحث داخل النص عبر `q`.
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/Query'
        - name: per_page
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: قائمة الرسائل
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Messages]
      summary: إرسال رسالة
      description: |
        يتطلب `messages.send`. أرسل JSON للنص أو `multipart/form-data` للوسائط.
        الحالة الأولية `sending` ثم تُحدَّث عبر الويب هوك إلى sent / delivered / read / failed.
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreMessageJsonRequest'
            example:
              type: text
              body: مرحباً، كيف يمكنني مساعدتك؟
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoreMessageMultipartRequest'
      responses:
        '201':
          description: تم إنشاء الرسالة وإرسالها للطابور
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CreatedEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum Token
    bridgeSecret:
      type: apiKey
      in: header
      name: X-Bridge-Secret
      description: السر المشترك بين Laravel وجسر WhatsApp (`WHATSAPP_BRIDGE_SECRET`)
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
...
    Query:
      name: q
      in: query
      description: بحث نصي
      schema:
        type: string
...
    ConversationId:
      name: conversation
      in: path
      required: true
      schema:
        type: integer
...
    TokenId:
      name: token
      in: path
      required: true
      schema:
        type: integer
  responses:
    MessageOk:
      description: نجاح
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    Unauthenticated:
      description: غير مصادق
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Unauthenticated.
...
    Forbidden:
      description: لا توجد صلاحية أو الحساب معطّل
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Forbidden.
...
    ValidationError:
      description: خطأ تحقق
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data: {}
    CreatedEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Created
        data: {}
    PaginatedEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 4
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 73
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ValidationErrorEnvelope:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.
...
    MessageType:
      type: string
      enum: [text, image, video, document, audio, sticker, location, contacts, unknown]
    MessageDirection:
      type: string
      enum: [inbound, outbound]
    MessageStatus:
      type: string
      enum: [sending, sent, delivered, read, failed]
...
    PermissionName:
      type: string
      enum:
        - customers.view
        - customers.create
        - customers.update
        - customers.delete
        - conversations.view
        - conversations.view_all
        - conversations.assign
        - conversations.close
        - conversations.delete
        - messages.view
        - messages.send
        - notes.create
        - tags.view
        - tags.manage
        - users.view
        - users.manage
        - whatsapp.manage
        - reports.view
        - audit.view
    Locale:
      type: string
      enum: [ar, en]
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        locale:
          $ref: '#/components/schemas/Locale'
        timezone:
          type: string
          example: Asia/Riyadh
        is_active:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        locale:
          $ref: '#/components/schemas/Locale'
        is_active:
          type: boolean
        is_online:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [admin, employee]
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionName'
        company:
          $ref: '#/components/schemas/Company'
        created_at:
          type: string
          format: date-time
...
    MediaFile:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: '#/components/schemas/MessageType'
        filename:
          type: string
          nullable: true
        mime_type:
          type: string
          nullable: true
        size:
          type: integer
        url:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
    Message:
      type: object
      properties:
        id:
          type: integer
        conversation_id:
          type: integer
        user_id:
          type: integer
          nullable: true
        direction:
          $ref: '#/components/schemas/MessageDirection'
        type:
          $ref: '#/components/schemas/MessageType'
        body:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/MessageStatus'
        error_message:
          type: string
          nullable: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaFile'
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        phone:
          type: string
          nullable: true
          maxLength: 30
        locale:
          $ref: '#/components/schemas/Locale'
        password:
          type: string
          minLength: 8
        password_confirmation:
          type: string
...
    StoreMessageJsonRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/MessageType'
        body:
          type: string
          maxLength: 4096
        caption:
          type: string
          maxLength: 1024
    StoreMessageMultipartRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/MessageType'
        body:
          type: string
        caption:
          type: string
        file:
          type: string
          format: binary
          description: حتى 16MB. صور، فيديو، مستندات، صوت.
...